Rename txid_set module to outpoint_set
