+++ src/bitcoin/outpoint_set.rs
